refactor(GetTagByIdQuery): implement strongly typed id and update namespace
