+++ src/ToDoManager.Application/Tags/Queries/GetTagById/GetTagByIdQueryHandler.cs
@@ -7,7 +7,7 @@ using ToDoManager.Domain.Tags.ValueObjects;
 
 namespace ToDoManager.Application.Tags.Queries.GetTagByIdQuery;
 
-public class GetTagByIdQueryHandler : IRequestHandler<GetTagByIdQuery, ErrorOr<DefaultTagResult>>
+public class GetTagByIdQueryHandler : IRequestHandler<GetTagById.GetTagByIdQuery, ErrorOr<DefaultTagResult>>
 {
 	private readonly ITagRepository _tagRepository;
 
@@ -16,7 +16,7 @@ public class GetTagByIdQueryHandler : IRequestHandler<GetTagByIdQuery, ErrorOr<D
 		_tagRepository = tagRepository;
 	}
 
-	public async Task<ErrorOr<DefaultTagResult>> Handle(GetTagByIdQuery request, CancellationToken cancellationToken)
+	public async Task<ErrorOr<DefaultTagResult>> Handle(GetTagById.GetTagByIdQuery request, CancellationToken cancellationToken)
 	{
 		var tag = await _tagRepository.GetByIdAsync(TagId.Create(request.Id));
